shell directory navigation (WIP)

+++ kernel/rt/atm.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-typedef uintptr_t k_rt_spnl_t;
+typedef uint32_t k_rt_spnl_t;
 
 typedef uintptr_t k_rt_cond_t;
